+++ test/integration/suite.go
@@ -200,6 +200,7 @@ func DefaultControllerConfigForTests() manager.Config {
 	cfg.ControllerName = "konghq.com/gateway-operator-integration-tests"
 	cfg.GatewayControllerEnabled = true
 	cfg.ControlPlaneControllerEnabled = true
+	cfg.ControlPlaneExtensionsControllerEnabled = true
 	cfg.DataPlaneControllerEnabled = true
 	cfg.DataPlaneBlueGreenControllerEnabled = bluegreenController
 	cfg.KongPluginInstallationControllerEnabled = true